% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{bench_fun}
\alias{bench_fun}
\title{benchmark returns calculator
eq_rets = zoo of equity returns 
fi_rets = zoo of fixed income returns
mkt_pct = weight of the equity
1 -mkt_pct is the weight of the fixed income}
\usage{
bench_fun(eq_rets, fi_rets, mkt_pct, start_date = "2005-10-31")
}
\description{
benchmark returns calculator
eq_rets = zoo of equity returns 
fi_rets = zoo of fixed income returns
mkt_pct = weight of the equity
1 -mkt_pct is the weight of the fixed income
}

