% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_bench_ETF}
\alias{contr_bench_ETF}
\title{contribution by benchmark for the ETF-based positions}
\usage{
contr_bench_ETF(start_date, end_date,
  eq_ret = composites$bench_rets_ETF_EQ[i],
  fi_ret = composites$bench_rets_ETF_FI[i], mkt_pct = composites$eq_pct[i],
  tickers = c(composites$bench_tickers_ETF_EQ[i],
  composites$bench_tickers_ETF_FI[i]), classer = "Broad_Category_Group")
}
\description{
contribution by benchmark for the ETF-based positions
}

