% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_lt}
\alias{contr_lt}
\title{long-term contribution}
\usage{
contr_lt(start_date, end_date, contr_data = paste(composites$STRAT_CODES[i],
  "S", sep = "_"))
}
\description{
long-term contribution
}

