% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_lt_class}
\alias{contr_lt_class}
\title{contribution by asset class}
\usage{
contr_lt_class(contr_tab, classer = "US_Broad_Asset_Class")
}
\description{
contribution by asset class
}

