% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_model}
\alias{contr_model}
\title{contribution for a model portfolio}
\usage{
contr_model(vsn, start_date, end_date, modeler, custom_rets = NULL)
}
\description{
contribution for a model portfolio
}

