% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_model_no_rebal}
\alias{contr_model_no_rebal}
\title{contribution for a model portfolio w/ no rebalances}
\usage{
contr_model_no_rebal(vsn, start_date, end_date, model, custom_rets = NULL)
}
\description{
contribution for a model portfolio w/ no rebalances
}

