% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_trades}
\alias{contr_trades}
\title{contribution by trades}
\usage{
contr_trades(start_date, end_date,
  contr_data = paste(composites$STRAT_CODES[i], "S", sep = "_"))
}
\description{
contribution by trades
}

