% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{model_fun}
\alias{model_fun}
\title{create return series from a model
vsn is for ETF, NAV, price or index-based returns; returns should be called "<type>_rets"
modeller is a zoo of the weights
custom_rets is for a non-loaded return series}
\usage{
model_fun(vsn, modeller, namer = "ret", custom_rets = NULL)
}
\description{
create return series from a model
vsn is for ETF, NAV, price or index-based returns; returns should be called "<type>_rets"
modeller is a zoo of the weights
custom_rets is for a non-loaded return series
}

