% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{model_fun_no_rebal}
\alias{model_fun_no_rebal}
\title{identical to model_fun, except does not rebalance}
\usage{
model_fun_no_rebal(vsn, model, namer = "ret", custom_rets = NULL)
}
\description{
identical to model_fun, except does not rebalance
}

