% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_shrink_factsheet}
\alias{perf_shrink_factsheet}
\title{APX to R conversion with EOD weights for factsheets}
\usage{
perf_shrink_factsheet(df, fees = "n")
}
\description{
APX to R conversion with EOD weights for factsheets
}

