% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{reader}
\alias{reader}
\title{specific to the .prf and .prb files from APX; clears the header}
\usage{
reader(df)
}
\description{
specific to the .prf and .prb files from APX; clears the header
}

