% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ret_roll_mo}
\alias{ret_roll_mo}
\title{trailing rolling returns
does NOT annualize long or short term}
\usage{
ret_roll_mo(rets, mos = 1)
}
\description{
trailing rolling returns
does NOT annualize long or short term
}

