% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ret_roll_mo_ann}
\alias{ret_roll_mo_ann}
\title{trailing rolling returns
annualize long term, leave short term unannualized}
\usage{
ret_roll_mo_ann(rets, mos = 1)
}
\description{
trailing rolling returns
annualize long term, leave short term unannualized
}

