% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ret_roll_mo_ann_all}
\alias{ret_roll_mo_ann_all}
\title{trailing rolling returns
annualize long and short term}
\usage{
ret_roll_mo_ann_all(rets, mos = 1)
}
\description{
trailing rolling returns
annualize long and short term
}

