% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{sd_roll}
\alias{sd_roll}
\title{trailing rolling standard deviations, for a daily series}
\usage{
sd_roll(rets, yrs = 1, tr = trade_days_US)
}
\description{
trailing rolling standard deviations, for a daily series
}

