% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{sd_roll_mo}
\alias{sd_roll_mo}
\title{trailing rolling standard deviations, for a monthly series}
\usage{
sd_roll_mo(rets, mos = 1)
}
\description{
trailing rolling standard deviations, for a monthly series
}

