% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{xyear_sd}
\alias{xyear_sd}
\title{variable year standard deviation for a daily series, end of month}
\usage{
xyear_sd(rets, yrs, tr = trade_days_US)
}
\description{
variable year standard deviation for a daily series, end of month
}

