% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{xyear_sd_m}
\alias{xyear_sd_m}
\title{variable year standard deviation for a daily series, monthly series}
\usage{
xyear_sd_m(rets, yrs)
}
\description{
variable year standard deviation for a daily series, monthly series
}

