% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{contr_bench_IDX}
\alias{contr_bench_IDX}
\title{contribution by benchmark for the Index-based positions}
\usage{
contr_bench_IDX(start_date, end_date,
  eq_ret = composites$bench_rets_IDX_EQ[i],
  fi_ret = composites$bench_rets_IDX_FI[i], mkt_pct = composites$eq_pct[i],
  tickers = c(composites$bench_tickers_IDX_EQ[i],
  composites$bench_tickers_IDX_FI[i]), classer = "Broad_Category_Group")
}
\description{
contribution by benchmark for the Index-based positions
}

